survey

+++ src/main/java/com/teamdefinex/dfxsurvey/application/impl/QuestionServiceImpl.java
@@ -1,31 +1,22 @@
 package com.teamdefinex.dfxsurvey.application.impl;
 
 import com.teamdefinex.dfxsurvey.application.QuestionService;
-import com.teamdefinex.dfxsurvey.data.AnswerRepository;
 import com.teamdefinex.dfxsurvey.data.QuestionRepository;
 import com.teamdefinex.dfxsurvey.data.SurveyRepository;
 import com.teamdefinex.dfxsurvey.domain.admin.QuestionType;
 import com.teamdefinex.dfxsurvey.domain.admin.Questions;
-import com.teamdefinex.dfxsurvey.domain.survey.Answer;
 import com.teamdefinex.dfxsurvey.domain.survey.Survey;
+import com.teamdefinex.dfxsurvey.domain.survey.SurveyStatus;
 import com.teamdefinex.dfxsurvey.dto.*;
 import com.teamdefinex.dfxsurvey.dto.result.Result;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.BeanUtils;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -43,11 +34,7 @@ public class QuestionServiceImpl implements QuestionService {
 
         Questions question = questionOptional.get();
 
-        QuestionSummaryResponseDTO response = new QuestionSummaryResponseDTO();
-        response.setId(question.getId().toString());
-        response.setQuestion(question.getQuestion());
-        response.setType(question.getType().name());
-        response.setOrder(question.getQuestionOrder());
+        QuestionSummaryResponseDTO response = createQuestionSummaryResponseDTO(question);
 
         return Result.success(response);
     }
@@ -61,18 +48,13 @@ public class QuestionServiceImpl implements QuestionService {
 
         Questions question = questionOptional.get();
 
-        QuestionDetailResponseDTO response = new QuestionDetailResponseDTO();
-        response.setId(question.getId().toString());
-        response.setQuestion(question.getQuestion());
-        response.setType(question.getType().name());
-        response.setOrder(question.getQuestionOrder());
-        response.setOptions(question.getOptions());
+        QuestionDetailResponseDTO response = createQuestionDetailResponseDTO(question);
 
         return Result.success(response);
     }
 
     @Transactional
-    public Result<Void> editQuestion(UUID id, EditQuestionRequestDTO request) {
+    public Result<QuestionDetailResponseDTO> editQuestion(UUID id, EditQuestionRequestDTO request) {
         Optional<Questions> questionOptional = questionRepository.findById(id);
 
         if(questionOptional.isEmpty()) {
@@ -81,13 +63,32 @@ public class QuestionServiceImpl implements QuestionService {
 
         Questions question = questionOptional.get();
 
+        if(question.getSurvey().getStatus() == SurveyStatus.SENT) {
+            return Result.failure("Survey already sent");
+        }
+
+        if(question.getSurvey().getStatus() == SurveyStatus.COMPLETED) {
+            return Result.failure("Survey already completed");
+        }
+
+        if(question.getSurvey().getStatus() == SurveyStatus.EXPIRED) {
+            return Result.failure("Survey expired");
+        }
+
         if(request.getQuestion() != null) question.setQuestion(request.getQuestion());
         if(request.getType() != null) question.setType(QuestionType.valueOf(request.getType()));
         if(request.getOrder() != null) question.setQuestionOrder(request.getOrder());
+        if (request.getOptions() != null && !request.getOptions().isEmpty()) {
+            question.getOptions().clear();
+            question.getOptions().addAll(request.getOptions());
+        }
+        question.setRequired(request.getRequired());
 
-        questionRepository.save(question);
+        Questions savedQuestion = questionRepository.save(question);
 
-        return Result.success(null);
+        QuestionDetailResponseDTO response = createQuestionDetailResponseDTO(savedQuestion);
+
+        return Result.success(response);
     }
 
     public Result<Void> deleteQuestion(UUID id) {
@@ -99,12 +100,24 @@ public class QuestionServiceImpl implements QuestionService {
 
         Questions question = questionOptional.get();
 
+        if(question.getSurvey().getStatus() == SurveyStatus.SENT) {
+            return Result.failure("Survey already sent");
+        }
+
+        if(question.getSurvey().getStatus() == SurveyStatus.COMPLETED) {
+            return Result.failure("Survey already completed");
+        }
+
+        if(question.getSurvey().getStatus() == SurveyStatus.EXPIRED) {
+            return Result.failure("Survey expired");
+        }
+
         questionRepository.delete(question);
 
         return Result.success(null);
     }
 
-    public Result<QuestionSummaryResponseDTO> duplicateQuestion(UUID id) {
+    public Result<QuestionDetailResponseDTO> duplicateQuestion(UUID id) {
         Optional<Questions> questionOptional = questionRepository.findById(id);
 
         if(questionOptional.isEmpty()) {
@@ -113,16 +126,28 @@ public class QuestionServiceImpl implements QuestionService {
 
         Questions question = questionOptional.get();
 
+        if(question.getSurvey().getStatus() == SurveyStatus.SENT) {
+            return Result.failure("Survey already sent");
+        }
+
+        if(question.getSurvey().getStatus() == SurveyStatus.COMPLETED) {
+            return Result.failure("Survey already completed");
+        }
+
+        if(question.getSurvey().getStatus() == SurveyStatus.EXPIRED) {
+            return Result.failure("Survey expired");
+        }
+
         Questions copy = question.duplicate();
 
         Questions savedQuestion = questionRepository.save(copy);
 
-        QuestionSummaryResponseDTO response = createQuestionSummaryResponseDTO(savedQuestion);
+        QuestionDetailResponseDTO response = createQuestionDetailResponseDTO(savedQuestion);
 
         return Result.success(response);
     }
 
-    public Result<QuestionSummaryResponseDTO> addQuestion(UUID surveyId, EditQuestionRequestDTO request, Authentication authentication) {
+    public Result<QuestionDetailResponseDTO> addQuestion(UUID surveyId, EditQuestionRequestDTO request, Authentication authentication) {
         Optional<Survey> surveyOptional = surveyRepository.findById(surveyId);
 
         if(surveyOptional.isEmpty()) {
@@ -137,19 +162,45 @@ public class QuestionServiceImpl implements QuestionService {
             return Result.failure("You are not owner of this Survey");
         }
 
+        if(survey.getStatus() == SurveyStatus.SENT) {
+            return Result.failure("Survey already sent");
+        }
+
+        if(survey.getStatus() == SurveyStatus.COMPLETED) {
+            return Result.failure("Survey already completed");
+        }
+
+        if(survey.getStatus() == SurveyStatus.EXPIRED) {
+            return Result.failure("Survey expired");
+        }
+
         Questions question = new Questions();
 
         question.setQuestion(request.getQuestion());
         question.setType(QuestionType.valueOf(request.getType()));
         question.setQuestionOrder(request.getOrder());
         question.setSurvey(survey);
+        question.setOptions(request.getOptions());
+        question.setRequired(request.getRequired());
 
         Questions savedQuestion = questionRepository.save(question);
 
-        QuestionSummaryResponseDTO response = createQuestionSummaryResponseDTO(savedQuestion);
+        QuestionDetailResponseDTO response = createQuestionDetailResponseDTO(savedQuestion);
 
         return Result.success(response);
 
+    }
+
+    private QuestionDetailResponseDTO createQuestionDetailResponseDTO(Questions question) {
+        QuestionDetailResponseDTO response = new QuestionDetailResponseDTO();
+        response.setId(question.getId().toString());
+        response.setQuestion(question.getQuestion());
+        response.setType(question.getType().name());
+        response.setOrder(question.getQuestionOrder());
+        response.setOptions(question.getOptions());
+        response.setRequired(question.getRequired());
+
+        return response;
     }
 
     private QuestionSummaryResponseDTO createQuestionSummaryResponseDTO(Questions question) {
@@ -158,6 +209,8 @@ public class QuestionServiceImpl implements QuestionService {
         response.setQuestion(question.getQuestion());
         response.setType(question.getType().name());
         response.setOrder(question.getQuestionOrder());
+        response.setRequired(question.getRequired());
+
         return response;
     }
